feature/shop-app-refactoring-redux-toolkit
